requirements.txt
dj-database-url
Django<3.1
django-debug-toolbar
django-rest-framework
djangorestframework
elasticsearch-dsl<8.0.0
factory-boy
psycopg2-binary